+++ packages/expo-modules-autolinking/android/expo-gradle-plugin/expo-autolinking-plugin/src/main/kotlin/expo/modules/plugin/ExpoAutolinkingPlugin.kt
@@ -112,6 +112,10 @@ open class ExpoAutolinkingPlugin : Plugin<Project> {
         (project.findProperty("expo.inlineModules.watchedDirectories")
           ?: emptyList<String>()).toString()
       )
+      it.kotlinPackageHeaderLength.set(
+        (project.findProperty("expo.inlineModules.kotlinPackageHeaderLength")
+          ?: "512").toString()
+      )
     }
   }