+++ packages/expo-modules-autolinking/src/commands/mirrorKotlinInlineModulesCommand.ts
@@ -6,6 +6,7 @@ import {
   createSymlinksToKotlinFiles,
   generateInlineModulesListFile,
 } from '../inlineModules/androidInlineModules';
+import type { KotlinPackageHeaderLength } from '../inlineModules/inlineModules';
 import { getMirrorStateObject } from '../inlineModules/inlineModules';
 import type { AutolinkingCommonArguments } from './autolinkingOptions';
 import { createAutolinkingOptionsLoader, registerAutolinkingArguments } from './autolinkingOptions';
@@ -14,6 +15,7 @@ interface MirrorKotlinInlineModulesCommandArguments extends AutolinkingCommonArg
   kotlinFilesMirrorDirectory: string;
   inlineModulesListDirectory: string;
   watchedDirectoriesSerialized: string;
+  kotlinPackageHeaderLength: string;
 }
 
 /**
@@ -36,17 +38,26 @@ export function mirrorKotlinInlineModulesCommand(cli: commander.CommanderStatic)
       '--watched-directories-serialized <watchedDirectories>',
       'JSON serialized watched directories array'
     )
+    .requiredOption(
+      '--kotlin-package-header-length <kotlinPackageHeaderLength>',
+      "Number of bytes read from the start of a Kotlin file to find its package declaration, or 'WHOLE_FILE'"
+    )
     .action(async (commandArguments: MirrorKotlinInlineModulesCommandArguments) => {
       const {
         kotlinFilesMirrorDirectory,
         inlineModulesListDirectory,
         watchedDirectoriesSerialized,
+        kotlinPackageHeaderLength: kotlinPackageHeaderLengthSerialized,
       } = commandArguments;
       const autolinkingOptionsLoader = createAutolinkingOptionsLoader({
         ...commandArguments,
       });
       const appRoot = await autolinkingOptionsLoader.getAppRoot();
       const watchedDirectories = JSON.parse(watchedDirectoriesSerialized);
+      const kotlinPackageHeaderLength: KotlinPackageHeaderLength =
+        kotlinPackageHeaderLengthSerialized === 'WHOLE_FILE'
+          ? 'WHOLE_FILE'
+          : Number(kotlinPackageHeaderLengthSerialized);
 
       if (
         !/.android./.test(kotlinFilesMirrorDirectory) ||
@@ -64,7 +75,11 @@ export function mirrorKotlinInlineModulesCommand(cli: commander.CommanderStatic)
         );
       }
 
-      const inlineModulesMirror = await getMirrorStateObject(watchedDirectories, appRoot);
+      const inlineModulesMirror = await getMirrorStateObject({
+        watchedDirectories,
+        appRoot,
+        kotlinPackageHeaderLength,
+      });
 
       const createMirrorStructurePromise = fs.promises
         .rm(kotlinFilesMirrorDirectory, { recursive: true, force: true })