+++ packages/expo-modules-autolinking/src/inlineModules/iosInlineModules.ts
@@ -1,18 +1,16 @@
 import type { ModuleIosConfig } from '../types';
+import type { InlineModulesScanOptions } from './inlineModules';
 import { getMirrorStateObject } from './inlineModules';
 
 export async function getIosInlineModulesClassNames(
-  watchedDirectories: string[],
-  appRoot: string
+  options: InlineModulesScanOptions
 ): Promise<ModuleIosConfig[]> {
-  return (await getMirrorStateObject(watchedDirectories, appRoot)).swiftModuleClassNames.map(
-    (className: string) => {
-      return {
-        class: className,
-        name: null,
-      };
-    }
-  );
+  return (await getMirrorStateObject(options)).swiftModuleClassNames.map((className: string) => {
+    return {
+      class: className,
+      name: null,
+    };
+  });
 }
 
 export function isTargetInInlineModulesTargets({